views.pyをスッキリさせたのと、招待ページの完成、通知ページの一部完成
