+++ the_project/mainpage/models.py
@@ -5,7 +5,6 @@ from django.contrib.auth.models import User
 # Create your models here.
 
 class Project(models.Model):
-    #project_id=models.CharField(max_length=100,primary_key=True)
     project_name=models.CharField(max_length=100)
 
     def __str__(self):
@@ -40,6 +39,7 @@ class Invite(models.Model):
     project_name = models.ForeignKey(Project, on_delete=models.CASCADE)
     invite_user = models.ForeignKey(User, on_delete=models.CASCADE, related_name='invite_user')
     invited_user = models.ForeignKey(User, on_delete=models.CASCADE, related_name='invited_user')
-    message = models.CharField(max_length=100, blank=True)  
+    message = models.CharField(max_length=100, blank=True)
 
-
+    def __str__(self):
+        return str(self.project_name)+","+str(self.invite_user)+","+str(self.message)
