+++ the_project/mainpage/programs/invite.py
@@ -1,0 +1,98 @@
+from django.shortcuts import render
+from django.views.generic import TemplateView
+# Create your views here.
+from django.contrib.auth.decorators import login_required
+from ..models import Project, Group, ProjectMember, Status, Invite
+from django.contrib.auth.models import User
+import pandas as pd
+from django.http import HttpResponse
+from ..forms import InviteForm
+
+#招待したいユーザーを検索 → 検索に合致するユーザーがいたらその人を選択 → messageを入力して招待を送信
+class InviteMembers(TemplateView):
+    def __init__(self):
+        self.params = {
+            "project_name" : "" ,
+            "form_serch_user" : InviteForm()["invited_user"],
+            "form_message" : InviteForm()["message"],
+            "degug" : [],
+            "message" : "",
+            "user" : "",
+            "msg":""
+        }
+    ################################################################################
+    ###処理で用いる関数たち
+    #すでにIDが存在しているかの判定
+    def exist_id(self, userid):
+        #IDで元DBを検索かけて中身が存在しなければFalseを返す
+        user_id = User.objects.filter(username=userid)
+        #これはクエリーセットを返しているけれど、クエリーセットは中身があればそれはTrue要素を含んでいるらしい
+        return user_id
+    #自分自身に招待を送った場合(同じ場合False、異なる場合True)
+    def owner(self, userid):
+        return userid != self.params["user"]
+    #################################################################################
+
+    def get(self, request, project_id):
+        self.params["user"] = request.GET["usr"]
+        self.params["project_name"] = project_id
+
+        # クエリパラメータ(?usr=<username>)で指定されたusernameをinvite_userに格納
+        self.invite_user = request.GET["usr"]
+        # クエリパラメータで指定されているプロジェクトネームを格納
+        self.invited_project = self.params["project_name"]
+
+        return render(request, 'mainpage/invite.html', self.params)
+
+    def post(self, request, project_id):
+        # フォームから入力された各値を格納
+        self.params["user"] = request.GET["usr"]
+        self.params["project_name"] = project_id
+        self.message = request.POST.get("message")
+        self.invited_user = request.POST.get("invited_user")
+        self.invite_user = self.params["user"]
+        self.invited_project = self.params["project_name"]
+
+        
+        
+
+        #登録できるか否か(Trueで登録可能、Falseで登録不可能)
+        exist_ornot = self.exist_id(self.invited_user)
+        owner_ornot = self.owner(self.invited_user)
+
+        recordable = exist_ornot and owner_ornot
+        #recordable = False
+        self.params["debug"] = owner_ornot
+
+        if recordable:
+            iv = Invite(
+                
+                invite_user=User.objects.get(username=self.invite_user), 
+                invited_user=User.objects.get(username=self.invited_user),
+                project_name=Project.objects.get(project_name=self.invited_project),
+                message=self.message
+                )
+            iv.save()
+            self.params["message"] = "招待メッセージを送りました。"
+            return render(request, "mainpage/invite_comp.html", self.params)
+        else:
+            if not(exist_ornot):
+                self.params["msg"] += "<br>そのようなユーザーIDは存在しません"
+
+            if not(owner_ornot):
+                self.params["msg"] += "<br>自分自身を招待することはできません"
+
+            self.params["msg"] = self.params["msg"][4:]
+            self.params["form_message"] = InviteForm(request.POST)["message"]
+            return render(request, "mainpage/invite.html", self.params)            
+        """
+        #データベースに格納
+        iv = Invite(
+            project_name=self.invited_project, 
+            invite_user=self.invite_user, 
+            invited_user=self.invited_user,
+            message=self.message
+            )
+        iv.save()
+        """
+
