+++ the_project/mainpage/templates/mainpage/grouppage.html
@@ -1,22 +1,21 @@
 {% load static %}
 <! dtype html >
 <html lang="ja">
-        <head>
-            <meta charset="utf-8">
-            <title>{{title}}</title>
-        </head>
-        <body>
-        <h1>グループページです</h1>
+    <head>
+        <meta charset="utf-8">
+        <title>{{title}}</title>
+    </head>
+    <body>
+        <h1>mainpageです</h1>
         {% if user.is_authenticated %}
 
         <a  href="{% url 'sign_in:sign_out' %}">Sign out</a>
         <p>ようこそ{{ userdata }}さん</p>
-
+        <a  href="/mainpage//notice">通知</a>
         <p>参加しているプロジェクト一覧</p>
         {%for item in project%}
         <a href="/mainpage/{{item.projectlist}}">{{item.projectlist}}</a><br>
         {% endfor %}
-
         <p>{{project_name}}に存在しているグループ一覧</p>
 
         <!--ユーザーが管理者かどうかを判別、管理者なら新しいグループを作成できるようにする-->
