+++ the_project/mainpage/templates/mainpage/invite_comp.html
@@ -1,1 +1,2 @@
 <h1>招待を送信しました</h1>>
+<a href="/mainpage/{{project_name}}">メインページに戻る</a>
