+++ the_project/mainpage/urls.py
@@ -10,6 +10,7 @@ urlpatterns = [
     path('<project_id>/', views.GroupList.as_view(), name='grouplist'),
     path('<project_id>/invite', views.InviteMembers.as_view(), name='invite'),
     path('<project_id>/create', views.CreateGroup.as_view(), name='create'),
+    path('/notice', views.Notify.as_view(), name='notify'),
     # グループページに飛ばすために、ここでgrouppageをインクルードしている
     path('<project_id>/grouppage', include('grouppage.urls'))
 ]
